refactored code and directory | added more comments
